Before Trying to String IF's
